Move tests to v1 package
