+++ app/src/test/java/com/mysugr/android/testing/v1/coroutine/TriggerAwaiterTest.kt
@@ -1,6 +1,6 @@
-package com.mysugr.android.testing.example.coroutine
+package com.mysugr.android.testing.v1.coroutine
 
-import com.mysugr.android.testing.example.appModuleTestingConfiguration
+import com.mysugr.android.testing.v1.example.appModuleTestingConfiguration
 import com.mysugr.sweetest.framework.base.BaseJUnitTest
 import com.mysugr.sweetest.framework.base.steps
 import com.mysugr.sweetest.framework.coroutine.testCoroutine
