+++ app/src/test/java/com/mysugr/android/testing/v1/example/FeatureRunner.kt
@@ -1,4 +1,4 @@
-package com.mysugr.android.testing.example
+package com.mysugr.android.testing.v1.example
 
 import cucumber.api.CucumberOptions
 import cucumber.api.junit.Cucumber
@@ -8,6 +8,6 @@ import org.junit.runner.RunWith
 @CucumberOptions(
     plugin = ["pretty"],
     features = ["src/test/resources/features"],
-    glue = ["com.mysugr.android.testing.example.feature", "com.mysugr.sweetest.framework.cucumber"]
+    glue = ["com.mysugr.android.testing.v1.example.feature", "com.mysugr.sweetest.framework.cucumber"]
 )
 class FeatureRunner
