+++ app/src/test/java/com/mysugr/android/testing/v1/example/net/BackendGatewaySteps.kt
@@ -1,7 +1,9 @@
-package com.mysugr.android.testing.example.net
+package com.mysugr.android.testing.v1.example.net
 
-import com.mysugr.android.testing.example.appModuleTestingConfiguration
-import com.mysugr.android.testing.example.feature.auth.UserSteps
+import com.mysugr.android.testing.v1.example.appModuleTestingConfiguration
+import com.mysugr.android.testing.v1.example.feature.auth.UserSteps
+import com.mysugr.android.testing.example.net.BackendGateway
+import com.mysugr.android.testing.example.net.UsernameOrPasswordWrongException
 import com.mysugr.android.testing.example.user.User
 import com.mysugr.sweetest.framework.base.BaseSteps
 import com.mysugr.sweetest.framework.base.dependency
