+++ app/src/test/java/com/mysugr/android/testing/v1/example/view/LoginIntegrationTest.kt
@@ -1,10 +1,11 @@
-package com.mysugr.android.testing.example.view
+package com.mysugr.android.testing.v1.example.view
 
-import com.mysugr.android.testing.example.appModuleTestingConfiguration
+import com.mysugr.android.testing.v1.example.appModuleTestingConfiguration
 import com.mysugr.android.testing.example.auth.AuthManager
-import com.mysugr.android.testing.example.feature.auth.UserSteps
-import com.mysugr.android.testing.example.net.BackendGatewaySteps
-import com.mysugr.android.testing.example.state.SessionStoreSteps
+import com.mysugr.android.testing.v1.example.feature.auth.UserSteps
+import com.mysugr.android.testing.v1.example.net.BackendGatewaySteps
+import com.mysugr.android.testing.v1.example.state.SessionStoreSteps
+import com.mysugr.android.testing.example.view.LoginViewModel
 import com.mysugr.sweetest.framework.base.BaseJUnitTest
 import com.mysugr.sweetest.framework.base.invoke
 import com.mysugr.sweetest.framework.base.steps
